requirements.txt
imaplib2>=3.6
email-validator>=2.0.0
requests>=2.31.0
flask==3.0.2
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
flask-login>=0.6.3
google-auth>=2.28.1
flask-session==0.6.0 